+++ dev_test_full_non_stop.py
@@ -101,7 +101,7 @@ def print_stream_data_from_stream_buffer(binance_websocket_api_manager):
     print("waiting 30 seconds, then we start flushing the stream_buffer")
     time.sleep(30)
     while True:
-        if binance_websocket_api_manager.stop_manager_request is not None:
+        if binance_websocket_api_manager.is_manager_stopping():
             exit(0)
         oldest_stream_data_from_stream_buffer = binance_websocket_api_manager.pop_stream_data_from_stream_buffer()
         if oldest_stream_data_from_stream_buffer is False: