+++ example_kline_1m.py
@@ -55,7 +55,7 @@ binance_get_kline_1m_bnbbtc = binance_websocket_api_manager.create_stream('kline
 
 def print_stream_data_from_stream_buffer(binance_websocket_api_manager):
     while True:
-        if binance_websocket_api_manager.stop_manager_request is not None:
+        if binance_websocket_api_manager.is_manager_stopping():
             exit(0)
         oldest_stream_data_from_stream_buffer = binance_websocket_api_manager.pop_stream_data_from_stream_buffer()
         if oldest_stream_data_from_stream_buffer is False: