+++ example_version_of_this_package.py
@@ -40,7 +40,8 @@ binance_websocket_api_manager = BinanceWebSocketApiManager()
 
 # get version of the used UNICORN Binance WebSocket API
 if binance_websocket_api_manager.is_update_availabe():
-    print("please upgrade to " + binance_websocket_api_manager.get_latest_version() + ", you are on", binance_websocket_api_manager.get_version())
+    print("please upgrade to " + binance_websocket_api_manager.get_latest_version() + ", you are on",
+          binance_websocket_api_manager.get_version())
 
     latest_release_info = binance_websocket_api_manager.get_latest_release_info()
     if latest_release_info: