+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.3.3.dev (development stage/unreleased)
+### Changed
+- get_monitoring_status_icinga(): update check
 
 ## 1.3.3
 ### Added