+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.3.4.dev (development stage/unreleased)
+### Added
+- TODO.md file
+### Fixed
+- is_update_available returns False if github API is not available.
 
 ## 1.3.4
 ### Changed